Move body utilities into their own file
