+++ packages/client-library-otel/src/types/index.ts
@@ -1,2 +1,8 @@
-export type { HonoLikeApp, HonoLikeEnv, HonoLikeFetch } from "./hono-types";
+export type {
+  HonoLikeApp,
+  HonoLikeEnv,
+  HonoLikeFetch,
+  GlobalResponse,
+  HonoResponse,
+} from "./hono-types";
 export { FetchFn } from "./runtime";
