+++ packages/client-library-otel/src/utils/request/headers.ts
@@ -1,5 +1,23 @@
 import type { FpResolvedConfig } from "../../config";
 import { getRedactedHeaders, getShouldTraceEverything } from "../../config";
+import type { GlobalResponse, HonoResponse } from "../../types";
+
+// There are so many different types of headers
+// and we want to support all of them so we can
+// use a single function to do it all
+type PossibleHeaders =
+  | Headers
+  | HonoResponse["headers"]
+  | GlobalResponse["headers"];
+
+export function headersToObject(headers: PossibleHeaders) {
+  const returnObject: Record<string, string> = {};
+  headers.forEach((value, key) => {
+    returnObject[key] = value;
+  });
+
+  return returnObject;
+}
 
 export function getSafeHeaderValue(
   key: string,
